fix: solve ThemeToggle icons not obeying theme
Before this commit, when you switch the website's theme to dark mode and refresh the page, the icon
of the ThemeToggle would revert to the light mode icon.
To fix this, instead of using useEffect to set the icon after the component is mounted, the easiest
solution is to render both icons at the same time and simply hide one of them accordingly to the
current theme.